SpnOverride: Fix of carrier display name and E-Plus/O2 national roaming
Patch proposed by user awl14 on XDA:
SpnOverride wil only apply the values in spn-conf.xml, if it is regarding
the network, but not if the SIM card provides a specific service provider
name. Further, improvements to recognize configured 'different' carriers,
which have been merged and where a respective config file exists.